feat: Update Yandex Metrika ID and improve header link
Updates the Yandex Metrika ID from 97931388 to 105075006 across the application to reflect the correct tracking ID.

Additionally, the header's logo link has been changed from a Telegram link to scroll to the top of the page for improved user experience and navigation. Scroll event listener is now passive and runs on mount. Minor text adjustments in Hero section. Privacy policy updated with a corrected name.